:truck: Rename load shim for call compatibility
This is a separate commit to help git track loader history across
renames. Please don't squash it.
